requirements.txt
adal==1.2.4
applicationinsights==0.11.9
argon2-cffi==20.1.0
attrs==20.2.0
autopep8==1.5.4
azure-common==1.1.25
azure-core==1.8.0
azure-graphrbac==0.61.1
azure-identity==1.2.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.13.0
azureml-core==1.13.0
azureml-datadrift==1.13.0
azureml-dataprep==2.0.7
azureml-dataprep-native==20.0.2
azureml-dataset-runtime==1.13.0
azureml-interpret==1.13.0
azureml-pipeline==1.13.0
azureml-pipeline-core==1.13.0
azureml-pipeline-steps==1.13.0
azureml-sdk==1.13.0
azureml-telemetry==1.13.0
azureml-train==1.13.0
azureml-train-automl-client==1.13.0.post1
azureml-train-core==1.13.0
azureml-train-restclients-hyperdrive==1.13.0
azureml-widgets==1.13.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cloudpickle==1.6.0
contextlib2==0.6.0.post1
cryptography==3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
docker==4.3.1
dotnetcore2==2.1.14
entrypoints==0.3
fusepy==3.0.1
idna==2.10
importlib-metadata==1.7.0
interpret-community==0.14.4
interpret-core==0.2.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.2
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
lightgbm==3.0.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
msal==1.5.0
msal-extensions==0.1.3
msrest==0.6.18
msrestazure==0.6.4
nbconvert==5.6.1
nbformat==5.0.7
ndg-httpsclient==0.5.1
notebook==6.1.4
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
portalocker==1.7.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
py4j==0.10.9
pyarrow==1.0.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.2
pyspark==3.0.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==3.1.2
Send2Trash==1.5.0
shap==0.34.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
zipp==3.1.0
